+++ packages/ai/CHANGELOG.md
@@ -35,6 +35,10 @@
 - Fixed resumed and forked Cursor sessions failing when their history came from a Responses-based provider such as Codex ([#9754](https://github.com/can1357/oh-my-pi/issues/9754)).
 - Fixed Cursor `composer-2.5` selections using the Fast variant instead of the Standard tier ([#9012](https://github.com/can1357/oh-my-pi/issues/9012)).
 
+### Added
+
+- Added Command Code API-key login with authenticated inference validation and mixed-wire request routing ([#9564](https://github.com/can1357/oh-my-pi/pull/9564) by [@himomohi](https://github.com/himomohi)).
+
 ## [18.0.6] - 2026-08-26
 
 ### Added